backend/requirements.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0

# HTTP client for Fortex API
httpx>=0.26.0
aiohttp>=3.9.0

# Browser automation
playwright>=1.41.0

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
python-dateutil>=2.8.2

# WebSocket
websockets>=12.0

# Event loop optimization for Windows (optional)
winloop>=0.1.0; sys_platform == 'win32'

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.0

# Development
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
